DP-674 Keep order of columns as defined in CSVLT
Since the underlying JSLT library filters out null fields by default,
the ordering of columns in the output CSV file can differ from the
order defined in the CSVLT transformation. This is because the CSV
printer outputs columns in the order it sees them, so if any of the
columns are empty on the first row, they will be moved to the back
(right) of the file.